💄 Modularie code
